Cloud_Funtions_GCP/etl_carga_maps/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-bigquery
pandas
numpy
pyarrow
fsspec
gcsfs